+++ appfl/algorithm/fedavg.py
@@ -4,7 +4,6 @@ log = logging.getLogger(__name__)
 
 from collections import OrderedDict
 from .algorithm import BaseServer, BaseClient
-from .misc import *
 
 import torch
 from torch.optim import *
@@ -16,24 +15,17 @@ import copy
 class FedAvgServer(BaseServer):
     def __init__(self, weights, model, num_clients, device, **kwargs):
         super(FedAvgServer, self).__init__(weights, model, num_clients, device)
-
         self.__dict__.update(kwargs)
-        self.num_clients = num_clients
-        self.weights = weights
-
-        self.primal_states = OrderedDict()
-        for i in range(num_clients):
-            self.primal_states[i] = OrderedDict()
 
     def update(self, local_states: OrderedDict):
 
-        """ Inputs """
+        """ Inputs for the global model update """
         global_state = OrderedDict()
-        primal_recover_from_local_states(self, local_states)
+        super(FedAvgServer, self).primal_recover_from_local_states(local_states)
         
 
-        """ Outputs """
-        for name, param in self.model.named_parameters():
+        """ global_state calculation """
+        for name, _ in self.model.named_parameters():
             tmp = 0.0
             for i in range(self.num_clients):
                 ## change device
@@ -43,6 +35,7 @@ class FedAvgServer(BaseServer):
 
             global_state[name] = tmp
 
+
         """ model update """
         self.model.load_state_dict(global_state)
 
@@ -50,10 +43,9 @@ class FedAvgServer(BaseServer):
 
 class FedAvgClient(BaseClient):
     def __init__(self, id, weight, model, dataloader, device, **kwargs):
-        super(FedAvgClient, self).__init__(id, weight, model, dataloader, device)
+        super(FedAvgClient, self).__init__(id, weight, model, dataloader, device)        
+        self.__dict__.update(kwargs)        
         self.loss_fn = CrossEntropyLoss()
-        self.__dict__.update(kwargs)
-        self.id = id
  
     def update(self):
 
@@ -62,7 +54,7 @@ class FedAvgClient(BaseClient):
 
         optimizer = eval(self.optim)(self.model.parameters(), **self.optim_args)
 
-        """ Inputs """                
+        """ Inputs for the local model update """         
         ## "global_state" from a server is already stored in 'self.model'
         
         """ Multiple local update """
@@ -78,23 +70,15 @@ class FedAvgClient(BaseClient):
                 loss.backward()
 
                 optimizer.step()
+        
+        self.primal_state = copy.deepcopy(self.model.state_dict())
 
-        """ Differential Privacy """
+        """ Differential Privacy  """
         if self.privacy == True:
-            # Note: Scale_value = Sensitivity_value / self.epsilon
-
-            Scale_value = self.scale_value
-
-            for name, param in self.model.named_parameters():
-                mean = torch.zeros_like(param.data)
-                scale = torch.zeros_like(param.data) + Scale_value
-                m = torch.distributions.laplace.Laplace(mean, scale)
-                param.data += m.sample()
-
+            super(FedAvgClient, self).laplace_mechanism_output_perturb()
+ 
         """ Update local_state """
         self.local_state = OrderedDict()
-        self.local_state["primal"] = OrderedDict()
-        for name, param in self.model.named_parameters():
-            self.local_state["primal"][name] = copy.deepcopy(param.data)
+        self.local_state["primal"] = copy.deepcopy(self.primal_state)
 
         return self.local_state
