+++ appfl/algorithm/iceadmm.py
@@ -4,7 +4,6 @@ log = logging.getLogger(__name__)
 
 from collections import OrderedDict
 from .algorithm import BaseServer, BaseClient
-from .misc import *
 
 import torch
 from torch.optim import *
@@ -16,32 +15,21 @@ import copy
 class ICEADMMServer(BaseServer):
     def __init__(self, weights, model, num_clients, device, **kwargs):
         super(ICEADMMServer, self).__init__(weights, model, num_clients, device)
-
-        self.__dict__.update(kwargs)
-        self.num_clients = num_clients
-        self.weights = weights
-
-        self.primal_states = OrderedDict()
-        self.dual_states = OrderedDict()
-        for i in range(num_clients):
-            self.primal_states[i] = OrderedDict()
-            self.dual_states[i] = OrderedDict()
-
-        self.penalty = OrderedDict()
-
+        self.__dict__.update(kwargs)        
+         
     def update(self, local_states: OrderedDict):
         
-        """ Inputs """
+        """ Inputs for the global model update"""
         global_state = self.model.state_dict()
-        primal_recover_from_local_states(self, local_states)
-        dual_recover_from_local_states(self, local_states)
-        penalty_recover_from_local_states(self, local_states)        
+        super(ICEADMMServer, self).primal_recover_from_local_states(local_states)
+        super(ICEADMMServer, self).dual_recover_from_local_states(local_states)
+        super(ICEADMMServer, self).penalty_recover_from_local_states(local_states)
 
         total_penalty = 0
         for i in range(self.num_clients):
             total_penalty += self.penalty[i]
 
-        """ Outputs """
+        """ global_state calculation """
         for name, param in self.model.named_parameters():
             tmp = 0.0
             for i in range(self.num_clients):
@@ -65,21 +53,19 @@ class ICEADMMServer(BaseServer):
 class ICEADMMClient(BaseClient):
     def __init__(self, id, weight, model, dataloader, device, **kwargs):
         super(ICEADMMClient, self).__init__(id, weight, model, dataloader, device)
-
-        self.loss_fn = CrossEntropyLoss()
         self.__dict__.update(kwargs)
-        self.id = id
-
-        self.model.to(device)
-        self.global_state = OrderedDict()
-        self.primal_state = OrderedDict()
-        self.dual_state = OrderedDict()
-        for name, param in model.named_parameters():            
+        self.loss_fn = CrossEntropyLoss()
+        
+        """ 
+        At initial, (1) primal_state = global_state, (2) dual_state = 0
+        """  
+        self.model.to(device)   
+        for name, param in model.named_parameters():                        
             self.primal_state[name] = param.data
             self.dual_state[name] = torch.zeros_like(param.data)
 
-        self.penalty = kwargs['penalty']      
-        self.proximity = kwargs['proximity']      
+        self.penalty = kwargs['init_penalty']      
+        self.proximity = kwargs['init_proximity']      
         self.residual = OrderedDict()
         self.residual['primal'] = 0
         self.residual['dual'] = 0            
@@ -91,10 +77,8 @@ class ICEADMMClient(BaseClient):
 
         optimizer = eval(self.optim)(self.model.parameters(), **self.optim_args)
 
-        """ Inputs """        
-        for name, param in self.model.named_parameters():
-            self.global_state[name] = copy.deepcopy(param.data)
-
+        """ Inputs for the local model update """         
+        global_state = copy.deepcopy(self.model.state_dict())        
 
         ## TODO: residual_calculation + adaptive_penalty
         ## Option 1: change penalty for every comm. round
@@ -125,32 +109,34 @@ class ICEADMMClient(BaseClient):
                 ## STEP: Update primal and dual
                 coefficient = 1
                 if self.coeff_grad == True:
-                    coefficient = self.weight * len(target) / len(self.dataloader.dataset)
-
-                iceadmm_step(self, coefficient)
-
-
-        """ Differential Privacy """
+                    coefficient = self.weight * len(target) / len(self.dataloader.dataset)                
+                iceadmm_step(self, coefficient, global_state)    
+ 
+        """ Differential Privacy  """
         if self.privacy == True:
-            # Note: Scale_value = Sensitivity_value / self.epsilon
-
-            Scale_value = self.scale_value
-
-            for name, param in self.model.named_parameters():
-                mean = torch.zeros_like(param.data)
-                scale = torch.zeros_like(param.data) + Scale_value
-                m = torch.distributions.laplace.Laplace(mean, scale)
-                self.primal_state[name] += m.sample()
+            super(ICEADMMClient, self).laplace_mechanism_output_perturb()
 
         """ Update local_state """
         self.local_state = OrderedDict()
-        self.local_state["primal"] = OrderedDict()
-        self.local_state["dual"] = OrderedDict()
-        for name, param in self.model.named_parameters():
-            self.local_state["primal"][name] = copy.deepcopy(self.primal_state[name])
-            self.local_state["dual"][name] = copy.deepcopy(self.dual_state[name])
-        
+        self.local_state["primal"] = copy.deepcopy(self.primal_state)
+        self.local_state["dual"] = copy.deepcopy(self.dual_state)                
         self.local_state["penalty"] = OrderedDict()
         self.local_state["penalty"][self.id] = self.penalty
 
         return self.local_state
+
+def iceadmm_step(self,coefficient,global_state):
+    for name, param in self.model.named_parameters():
+
+        grad = param.grad * coefficient
+        ## Update primal
+        self.primal_state[name] = self.primal_state[name] - (
+            self.penalty * (self.primal_state[name] - global_state[name])
+            + grad
+            + self.dual_state[name]
+        ) / (self.weight * self.proximity + self.penalty)
+        ## Update dual
+        self.dual_state[name] = self.dual_state[name] + self.penalty * (
+            self.primal_state[name] - global_state[name]
+        )
+
