+++ appfl/algorithm/iiadmm.py
@@ -4,7 +4,6 @@ log = logging.getLogger(__name__)
 
 from collections import OrderedDict
 from .algorithm import BaseServer, BaseClient
-from .misc import *
 
 import torch
 from torch.optim import *
@@ -19,31 +18,23 @@ class IIADMMServer(BaseServer):
         super(IIADMMServer, self).__init__(weights, model, num_clients, device)
 
         self.__dict__.update(kwargs)
-        self.num_clients = num_clients
-        self.weights = weights
-
-        self.primal_states = OrderedDict()
-        self.dual_states = OrderedDict()
-        for i in range(num_clients):
-            self.primal_states[i] = OrderedDict()
-            self.dual_states[i] = OrderedDict()
-            ## dual computation at local
-            for name, param in model.named_parameters():
-                self.dual_states[i][name] = torch.zeros_like(param.data)
         
-        self.penalty = OrderedDict()
+        """ 
+        At initial, dual_state = 0
+        """  
+        for i in range(num_clients):                        
+            for name, param in model.named_parameters():
+                self.dual_states[i][name] = torch.zeros_like(param.data)               
 
     def update(self, local_states: OrderedDict):
     
-        """ Inputs """
+        """ Inputs for the global model update"""
         global_state = self.model.state_dict()
-        primal_recover_from_local_states(self, local_states)        
-        penalty_recover_from_local_states(self, local_states)        
+        super(IIADMMServer, self).primal_recover_from_local_states(local_states)
+        super(IIADMMServer, self).penalty_recover_from_local_states(local_states)        
 
-        # for i in range(self.num_clients):
-        #     print("self.penalty[",i,"]=", self.penalty[i])
 
-        """ Outputs """
+        """ global_state calculation """
         for name, param in self.model.named_parameters():
             tmp = 0.0
             for i in range(self.num_clients):
@@ -72,20 +63,18 @@ class IIADMMServer(BaseServer):
 class IIADMMClient(BaseClient):
     def __init__(self, id, weight, model, dataloader, device, **kwargs):
         super(IIADMMClient, self).__init__(id, weight, model, dataloader, device)
-
-        self.loss_fn = CrossEntropyLoss()
         self.__dict__.update(kwargs)
-        self.id = id
-
-        self.model.to(device)                
-        self.global_state = OrderedDict()
-        self.primal_state = OrderedDict()
-        self.dual_state = OrderedDict()
-        for name, param in model.named_parameters():            
+        self.loss_fn = CrossEntropyLoss()
+                
+        """ 
+        At initial, (1) primal_state = global_state, (2) dual_state = 0
+        """  
+        self.model.to(device)  
+        for name, param in model.named_parameters():                        
             self.primal_state[name] = param.data
             self.dual_state[name] = torch.zeros_like(param.data)
         
-        self.penalty = kwargs['penalty']        
+        self.penalty = kwargs['init_penalty']        
         self.residual = OrderedDict()
         self.residual['primal'] = 0
         self.residual['dual'] = 0
@@ -97,9 +86,8 @@ class IIADMMClient(BaseClient):
 
         optimizer = eval(self.optim)(self.model.parameters(), **self.optim_args)
 
-        """ Inputs """        
-        for name, param in self.model.named_parameters():
-            self.global_state[name] = copy.deepcopy(param.data)
+        """ Inputs for the local model update """           
+        global_state = copy.deepcopy(self.model.state_dict())
         
         
         ## TODO: residual_calculation + adaptive_penalty
@@ -138,35 +126,68 @@ class IIADMMClient(BaseClient):
 
                     coefficient = self.weight  ## NOTE: BATCH + FEMNIST, rho=0.07
 
-                iiadmm_step(self, coefficient, optimizer)
+                iiadmm_step(self, coefficient, global_state, optimizer)
  
 
         ## Update dual
         for name, param in self.model.named_parameters():
             self.dual_state[name] = self.dual_state[name] + self.penalty * (
-                self.global_state[name] - self.primal_state[name]
+                global_state[name] - self.primal_state[name]
             )
 
-        """ Differential Privacy """
+        """ Differential Privacy  """
         if self.privacy == True:
-            # Note: Scale_value = Sensitivity_value / self.epsilon
-
-            Scale_value = self.scale_value
-
-            for name, param in self.model.named_parameters():
-                mean = torch.zeros_like(param.data)
-                scale = torch.zeros_like(param.data) + Scale_value
-                m = torch.distributions.laplace.Laplace(mean, scale)
-                self.primal_state[name] += m.sample()
+            super(IIADMMClient, self).laplace_mechanism_output_perturb()
 
         """ Update local_state """
-        self.local_state = OrderedDict()
-        
-        self.local_state["primal"] = OrderedDict()        
-        for name, param in self.model.named_parameters():
-            self.local_state["primal"][name] = copy.deepcopy(self.primal_state[name])
-        
+        self.local_state = OrderedDict()                
+        self.local_state["primal"] = copy.deepcopy(self.primal_state)                        
         self.local_state["penalty"] = OrderedDict()
         self.local_state["penalty"][self.id] = self.penalty
 
         return self.local_state
+
+
+def optimizer_setting(self):
+    momentum = 0
+    if "momentum" in self.optim_args.keys():
+        momentum = self.optim_args.momentum
+    weight_decay = 0
+    if "weight_decay" in self.optim_args.keys():
+        weight_decay = self.optim_args.weight_decay
+    dampening = 0
+    if "dampening" in self.optim_args.keys():
+        dampening = self.optim_args.dampening
+    nesterov = False
+
+    return momentum, weight_decay, dampening, nesterov
+
+
+def iiadmm_step(self, coefficient, global_state, optimizer):
+
+    momentum, weight_decay, dampening, nesterov = optimizer_setting(self)
+
+    for name, param in self.model.named_parameters():
+
+        grad = copy.deepcopy(param.grad * coefficient)
+
+        if weight_decay != 0:
+            grad.add_(weight_decay, self.primal_state[name])
+        if momentum != 0:
+            param_state = optimizer.state[param]
+            if "momentum_buffer" not in param_state:
+                buf = param_state["momentum_buffer"] = grad.clone()
+            else:
+                buf = param_state["momentum_buffer"]
+                buf.mul_(momentum).add_(1 - dampening, grad)
+            if nesterov:
+                grad = self.grad[name].add(momentum, buf)
+            else:
+                grad = buf
+
+        ## Update primal
+        self.primal_state[name] = global_state[name] + (1.0 / self.penalty) * (
+            self.dual_state[name] - grad
+        )
+
+
