+++ appfl/misc/utils.py
@@ -54,7 +54,7 @@ def print_write_result_title(cfg: DictConfig, DataSet_name: str):
             cfg.fed.args.coeff_grad,
             cfg.fed.args.privacy, 
             cfg.fed.args.scale_value,
-            cfg.fed.args.penalty,
+            cfg.fed.args.init_penalty,
         )
     if cfg.fed.type == "iceadmm":
         filename = "Result_%s_%s_Batch_%s_AccumGrad_%s_Privacy_%s_ScaleVal_%s(rho=%s,prox=%s)" % (
@@ -64,8 +64,8 @@ def print_write_result_title(cfg: DictConfig, DataSet_name: str):
             cfg.fed.args.accum_grad,
             cfg.fed.args.privacy, 
             cfg.fed.args.scale_value,
-            cfg.fed.args.penalty,
-            cfg.fed.args.proximity
+            cfg.fed.args.init_penalty,
+            cfg.fed.args.init_proximity
         )        
 
     file_ext = ".txt"
